+++ dramatiq/results/backend.py
@@ -1,7 +1,9 @@
 import hashlib
 import time
+import typing
 
 from ..common import compute_backoff, q_name
+from ..encoder import Encoder, JSONEncoder
 from .errors import ResultTimeout, ResultMissing
 
 #: The default timeout for blocking get operations in milliseconds.
@@ -13,19 +15,35 @@ BACKOFF_FACTOR = 100
 #: Canary value that is returned when a result hasn't been set yet.
 Missing = type("Missing", (object,), {})()
 
+#: A type alias representing backend results.
+Result = typing.Any
+
+#: A union representing a Result that may or may not be there.
+MResult = typing.Union[type(Missing), Result]
+
 
 class ResultBackend:
     """ABC for result backends.
+
+    Parameters:
+      namespace(str): The logical namespace under which the data
+        should be stored.
+      encoder(Encoder): The encoder to use when storing and retrieving
+        result data.  Defaults to :class:`.JSONEncoder`.
     """
 
-    def get_result(self, message, *, block=False, timeout=None):
+    def __init__(self, *, namespace: str="dramatiq-results", encoder: Encoder=None):
+        self.namespace = namespace
+        self.encoder = encoder or JSONEncoder()
+
+    def get_result(self, message: "Message", *, block: bool=False, timeout: int=None) -> Result:
         """Get a result from the backend.
 
         Parameters:
           message(Message)
           block(bool): Whether or not to block until a result is set.
           timeout(int): The maximum amount of time, in ms, to wait for
-            a result when block is True.
+            a result when block is True.  Defaults to 10 seconds.
 
         Raises:
           ResultMissing: When block is False and the result isn't set.
@@ -40,8 +58,8 @@ class ResultBackend:
 
         attempts = 0
         while True:
-            data = self._get(message_key)
-            if data is Missing and block:
+            result = self._get(message_key)
+            if result is Missing and block:
                 attempts, delay = compute_backoff(attempts, factor=BACKOFF_FACTOR)
                 delay /= 1000
                 if time.monotonic() + delay > end_time:
@@ -50,13 +68,13 @@ class ResultBackend:
                 time.sleep(delay)
                 continue
 
-            elif data is Missing:
+            elif result is Missing:
                 raise ResultMissing(message)
 
             else:
-                return data
+                return result
 
-    def store_result(self, message, result, ttl):
+    def store_result(self, message: "Message", result: Result, ttl: int) -> None:
         """Store a result in the backend.
 
         Parameters:
@@ -68,7 +86,7 @@ class ResultBackend:
         message_key = self.build_message_key(message)
         return self._store(message_key, result, ttl)
 
-    def build_message_key(self, message):
+    def build_message_key(self, message: "Message") -> str:
         """Given a message, return its globally-unique key.
 
         Parameters:
@@ -85,12 +103,20 @@ class ResultBackend:
         }
         return hashlib.md5(message_key.encode("utf-8")).hexdigest()
 
-    def _get(self, message_key):  # pragma: no cover
-        """Get a result from the backend.
+    def _get(self, message_key: str) -> MResult:  # pragma: no cover
+        """Get a result from the backend.  Subclasses may implement
+        this method if they want to use the default, polling,
+        implementation of get_result.
         """
-        raise NotImplementedError
+        raise NotImplementedError("%(classname)r does not implement _get()" % {
+            "classname": type(self).__name__,
+        })
 
-    def _store(self, message_key, result, ttl):  # pragma: no cover
-        """Store a result in the backend.
+    def _store(self, message_key: str, result: Result, ttl: int) -> None:  # pragma: no cover
+        """Store a result in the backend.  Subclasses may implement
+        this method if they want to use the default implementation of
+        set_result.
         """
-        raise NotImplementedError
+        raise NotImplementedError("%(classname)r does not implement _store()" % {
+            "classname": type(self).__name__,
+        })