+++ dramatiq/results/backends/redis.py
@@ -1,14 +1,16 @@
-import json
 import redis
 
-from ..backend import ResultBackend, Missing
+from ..backend import DEFAULT_TIMEOUT, ResultBackend, ResultMissing, ResultTimeout
 
 
 class RedisBackend(ResultBackend):
-    """A result backend for Redis_.
+    """A result backend for Redis_.  This is the recommended result
+    backend as waiting for a result is resource efficient.
 
     Parameters:
       namespace(str): A string with which to prefix result keys.
+      encoder(Encoder): The encoder to use when storing and retrieving
+        result data.  Defaults to :class:`.JSONEncoder`.
       client(StrictRedis): An optional client.  If this is passed,
         then all other parameters are ignored.
       \**parameters(dict): Connection parameters are passed directly
@@ -17,17 +19,51 @@ class RedisBackend(ResultBackend):
     .. _redis: https://redis.io
     """
 
-    def __init__(self, *, namespace="dramatiq-results", client=None, **parameters):
-        self.namespace = namespace
+    def __init__(self, *, namespace="dramatiq-results", encoder=None, client=None, **parameters):
+        super().__init__(namespace=namespace, encoder=encoder)
         self.client = client or redis.StrictRedis(**parameters)
 
-    def _get(self, message_key):
-        data = self.client.get(message_key)
-        if data is not None:
-            return json.loads(data.decode("utf-8"))
-        return Missing
+    def get_result(self, message, *, block=False, timeout=None):
+        """Get a result from the backend.
+
+        Warning:
+          Sub-second timeouts are not respected by this backend.
+
+        Parameters:
+          message(Message)
+          block(bool): Whether or not to block until a result is set.
+          timeout(int): The maximum amount of time, in ms, to wait for
+            a result when block is True.  Defaults to 10 seconds.
+
+        Raises:
+          ResultMissing: When block is False and the result isn't set.
+          ResultTimeout: When waiting for a result times out.
+
+        Returns:
+          object: The result.
+        """
+        timeout = timeout or DEFAULT_TIMEOUT
+        message_key = self.build_message_key(message)
+        if block:
+            timeout = int(timeout / 1000)
+            if timeout == 0:
+                data = self.client.rpoplpush(message_key, message_key)
+            else:
+                data = self.client.brpoplpush(message_key, message_key, timeout)
+
+            if data is None:
+                raise ResultTimeout(message)
+
+        else:
+            data = self.client.lindex(message_key, 0)
+            if data is None:
+                raise ResultMissing(message)
+
+        return self.encoder.decode(data)
 
     def _store(self, message_key, result, ttl):
-        result_data = json.dumps(result, separators=(",", ":")).encode("utf-8")
-        expiration = int(ttl / 1000)
-        self.client.setex(message_key, expiration, result_data)
+        with self.client.pipeline() as pipe:
+            pipe.delete(message_key)
+            pipe.lpush(message_key, self.encoder.encode(result))
+            pipe.pexpire(message_key, ttl)
+            pipe.execute()